Update fmi_util.pyx
Added +1 to lengths of list in order to make sure we have an empty character in the generated string for the default variable time. This is because the function convert_array_names_list_names_int loops through the characters in each variable name and breaks when it finds an empty character, i.e. chr(0). This will resolve issue with the result key for 'time' being displayed as 'tim' instead of desired 'time'. If we don't add +1 we end up with breaking the loop too early when we simulate with options["filter"] set to any value that would exclude all other variables with names longer than 4 characters.